feature:点选图形验证码代码优化调整；

+++ captcha-spring-boot-starter/src/main/java/com/emily/captcha/otp/OtpSecretGenerator.java
@@ -124,7 +124,7 @@ public class OtpSecretGenerator {
      * @return OTP Auth URI
      */
     public static String generateOtpAuthUri(String secret, String account, String issuer,
-                                            OtpHashAlgorithm algorithm, int digits, int period) {
+                                            OtpHashAlgorithm algorithm, int digits, long period) {
         return String.format("otpauth://totp/%s:%s?secret=%s&issuer=%s&algorithm=%s&digits=%d&period=%d",
                 encodeUriComponent(issuer), encodeUriComponent(account),
                 secret, encodeUriComponent(issuer), algorithm.getUriAlgorithm(), digits, period);
